Moving generic function to other location
